Pack each formatting run into one long in a BIGINT array.
Store chat_entry.formatting as BIGINT[] again. Offset and count use 16 bits
each (Minecraft chat is at most 65535 characters) and format uses the high
32 bits, so one long holds a whole range.

+++ src/main/java/me/wolfii/allthelogs/client/ComponentFormatting.java
@@ -25,7 +25,7 @@ public final class ComponentFormatting {
         message.visit((style, string) -> {
             FormattingCodes.Parsed parsed = FormattingCodes.parse(string, pack(style));
             text.append(parsed.text());
-            int[] packed = parsed.formatting();
+            long[] packed = parsed.formatting();
             for (int i = 0; i < parsed.text().length(); i++) {
                 perChar.add(PackedFormatting.at(packed, i));
             }
