+++ src/main/java/me/wolfii/allthelogs/client/ui/text/MessageText.java
@@ -140,7 +140,7 @@ public final class MessageText {
         String text = full.substring(start, end);
         if (text.isEmpty()) return Component.empty();
         boolean interpret = VisualMessage.interpretEscapes(row.chatLog());
-        int[] formatting = VisualMessage.remapFormatting(row.entry().message(), row.entry().formatting(), interpret);
+        long[] formatting = VisualMessage.remapFormatting(row.entry().message(), row.entry().formatting(), interpret);
         MutableComponent result = Component.empty();
         int runStart = 0;
         int runFormat = PackedFormatting.at(formatting, start);
@@ -162,7 +162,7 @@ public final class MessageText {
      * Chat colour with match highlight, context dimming, and {@code \n} darkening multiplied in that order.
      */
     public static int stackedColor(DisplayRow row, int index, boolean interpretEscapes) {
-        int[] formatting = VisualMessage.remapFormatting(row.entry().message(), row.entry().formatting(),
+        long[] formatting = VisualMessage.remapFormatting(row.entry().message(), row.entry().formatting(),
             interpretEscapes);
         return stackedColor(row, index, interpretEscapes, PackedFormatting.at(formatting, index));
     }
