+++ src/main/java/me/wolfii/allthelogs/data/parse/ParsedLog.java
@@ -29,7 +29,7 @@ public record ParsedLog(
      * @param message     everything after {@code [CHAT] }, without legacy {@code §} codes
      * @param formatting  packed {@code (offset, count, format)} triples, or {@code null}
      */
-    public record Entry(LocalTime time, String message, int[] formatting) {
+    public record Entry(LocalTime time, String message, long[] formatting) {
         public Entry(LocalTime time, String message) {
             this(time, message, null);
         }
